Don't try to delete the same hostname twice
The previous fix (to cover two different hostnames) introduced a
regression where the same hostname would get deleted twice.

+++ OpenStack-Rabbit-Consumer/rabbit_consumer/message_consumer.py
@@ -76,10 +76,13 @@ def delete_machine(
 
     hostname = aq_api.search_host_by_machine(machine_name)
     machine_details = aq_api.get_machine_details(machine_name)
+    # We have to clean-up all the interfaces and addresses first
     if hostname:
-        # We have to clean-up all the interfaces and addresses first
-        logger.info("Host exists for %s. Deleting old", hostname)
-        aq_api.delete_host(hostname)
+        if aq_api.check_host_exists(hostname):
+            # This is a different hostname to the one we have in the message
+            # so, we need to delete it
+            logger.info("Host exists for %s. Deleting old", hostname)
+            aq_api.delete_host(hostname)
 
         # First delete the interfaces
         ipv4_address = socket.gethostbyname(hostname)
